requirements.txt
torch
torchvision
datasets
torchmetrics
matplotlib
Pillow
tqdm
